[CST-5269] move iiif cache loggers into dspace-api progect
