+++ dspace-api/src/main/java/org/dspace/iiif/logger/CanvasCacheLogger.java
@@ -5,7 +5,7 @@
  *
  * http://www.dspace.org/license/
  */
-package org.dspace.app.rest.cache;
+package org.dspace.iiif.logger;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
